+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+  <<<<<<< fix-kubernetes-context
+## [Unreleased]
+### Added
+- `-U --uppercase` flag to convert `lower-case` keys in credstash to `LOWER_CASE` secrets
+- `pushall` and `daemonall` commands
+  =======
 ## 2.0.2 - 2019-05-11
 ### Changes
 - added support for loading incluster kubernetes config
@@ -20,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `pushall` and `daemonall` commands
 - various bug fixes
 - removed `inject` command
+  >>>>>>> envfrom
 
 ## [1.0.0] - 2018-04-03
 ### Added
@@ -27,4 +34,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `-l --lowercase` flag to enable DNS_SUBDOMAIN key interpolation
 
 ### Changed
+  <<<<<<< fix-kubernetes-context
 - keys are no longer converted to DNS_SUBDOMAIN by default
+  =======
+- keys are no longer converted to DNS_SUBDOMAIN by default
+  >>>>>>> envfrom